Use`UShort` instead of `UInt` to save 2 bytes in memory per value

+++ app/src/test/java/com/sebiai/glyphport/TestingUtils.kt
@@ -13,7 +13,7 @@ object CompositionLightDataLoader {
             .map { it.trim().removeSuffix(",") }
             .filter { it.isNotEmpty() }
             .map { it.split(',') }
-            .map { it.map { brightness -> brightness.toUInt() } }
+            .map { it.map { brightness -> brightness.toUShort() } }
 
         return CompositionLightData(lightData)
     }
